requirements.txt
streamlit
requests
urllib3
google-generativeai
python-dotenv